[focus] Only redirect on external URLs if really necessary (https://github.com/mozilla-mobile/focus-android/issues/281 and https://github.com/mozilla-mobile/focus-android/issues/282)
Since we now verify URLs when first loading a URL,
we would end up in a loop when we had an unsupported URL
(e.g. foo:bar....), and we'd then try to load that URL again in our
external URL handling code. The correct behaviour is to just stop
loading, and let the error handling code take over.

(This of course makes the logic more complicated, which will be fixed
 in an upcoming commit.)