[focus] Ensure BrowserFragment is visible before querying state (https://github.com/mozilla-mobile/focus-android/issues/287)
It seems we can crash if BrowserFragment has been added, but it's
View isn't created yet (in this case webView will be null, so
any operations calling webView's methods will crash).

We shouldn't actually be asking about BrowserFragment's history unless
it's actually on screen, so adding this check here is useful in general.
(We should also fix BrowserFragment to not query webView unless it actually
 exists, which will happen in a later commit.)