[focus] Revert "Ensure SearchEngine data is loaded before loading prefs (https://github.com/mozilla-mobile/focus-android/issues/405)"
This was a quick fix to avoid hanging. Instead we should make sure that
SearchEnginePreference doesn't try to load searche engine data
until it's actually being shown (at which point search engine data will
be loaded anyway).

We don't need to revert this commit, but we do want to remove the
now-obsolete comment, and we might as well cleanly revert.

This reverts commit 4375da46d2a96b02b9219118fcd6279f9ffa5c0a.