[focus] Closes https://github.com/mozilla-mobile/focus-android/issues/838: Dismiss HomeFragment's PopupMenu onDetach.
One undesirable effect of this change: this issue only shows itself when we
.replace the HomeFragment, which happens when we change locales, whose code
only gets called from `onResume`. This means the menu is *briefly* visible when
returning to the app when following the STR in the bug. This is ugly but given
how infrequently I expect this to happen, I think it's fine.

The solution would be to additionally dismiss the PopupMenu in onStop, but:

1. I'm concerned about unforseen side effects from that change
2. For correctness, we'd have to dismiss the PopupMenu in both `onDetach` and
`onStop` which is a little confusing from a code perspective - better to keep
it clean, I think!