[focus] Restructure BrowserFragment layout to cast toolbar shadow and fix UI issues. (https://github.com/mozilla-mobile/focus-android/issues/1108, https://github.com/mozilla-mobile/focus-android/issues/785)
With this change we'll always cast a toolbar shadow (https://github.com/mozilla-mobile/focus-android/issues/1108) and prevent the loading bar from
appearing before the toolbar (https://github.com/mozilla-mobile/focus-android/issues/785). Additionally it adds an alpha animation to the toolbar
content when scrolling so that it does not clash with the status bar.