[focus] [detekt] Disable WildcardImport rule.
We do not allow wildcard imports for Java code (via checkstyle). However for
Kotlin code Android Studio / IntelliJ automatically uses wildcard imports if
5 or more imports from the same package are used.

As long as this is the default in the IDE I do not think it's viable to fight
against the default setting and ask every developer to setup the IDE
differently.