[focus] Closes https://github.com/mozilla-mobile/focus-android/issues/1875: Remove cookies on new session.
This is the simplest fix that aims to avoid unwanted side effects: this means
there may be other hidden issues we're not fixing here. We should try to
comprehensively fix this bug in issue https://github.com/mozilla-mobile/focus-android/issues/1569.

A better fix with potential side effects would be forcing WebView.cleanup to
run on a new browser session, not just deleting cookies.