[fenix] Try to fix build bustage from https://github.com/mozilla-mobile/fenix/issues/1321. (https://github.com/mozilla-mobile/fenix/pull/1538)
This just looks wrong.  I'm concerned that this was green as a PR,
because it suggests that we were signing with an incorrect
`worker_type` there; but here we are -- let's see if it addresses the
issue.  If not, we'll back the whole of https://github.com/mozilla-mobile/fenix/issues/1321 out.