[fenix] Generate the `activation` ping and send it with Glean (https://github.com/mozilla-mobile/fenix/pull/1707)
This fetches the Google Advertising ID, salts it and
then applies hashing before sending a ping with it,
at startup. Hashing and salting are used in order
to prevent ourselves to correlate advertising IDs
from the same user running different products we
own off a single device. We will never send the
client_id and the Google Advertising ID in the same
ping.