[fenix] Closes https://github.com/mozilla-mobile/fenix/issues/2770: Allow receiving tabs from FxA devices (https://github.com/mozilla-mobile/fenix/pull/2772)
* Closes https://github.com/mozilla-mobile/fenix/issues/2770: Allow receiving tabs from FxA devices

Now that we're on a-c 0.54, we can land this since it supports device
capability migration.

This patch adds a SEND_TAB device capability, making Fenix a valid target
in the Send Tab device list on Desktop Firefox.

Additionally, it adds a notification manager which manages notification
channels and knows how to display "received tabs" notifications".

* Post: remove unusued test file that's causing issues