[fenix] Fix 4827 - Use ktx doOnPreDraw() to restore HomeFragment's layout
The previous solution would result in a crash because the passed in
viewTreeObserver that would trigger onPreDraw would be invalid.
The proposed solution is simpler and ensures we'll always use the right
viewTreeObserver.

`FragmentPreDrawManager` is general enough that can be used by other Fragments
also, so I've added it to the `utils` package.