[components] Part 1: Introduce re-usable SyncTelemetry
We will need to record sync telemetry from the sync manager. We already
have telemetry related code, but at the DB Connection level.

This patch refactors this code into a singleton object that may be
used from within any context in which there's a ping object available that
needs to be processed.