[fenix] For https://github.com/mozilla-mobile/fenix/issues/3584 - Check if menu is open before showing a new one (https://github.com/mozilla-mobile/fenix/pull/5674)
Could have implemented this check (if menu is showing) inside the show() method
of BrowserMenu but this would mean the client (us) would go to the process of
building a new menu and then trying to have it displayed only for this to be
ignored by BrowserMenu in a somewhat opaque way.
Having this check done as soon as possible offers us full control and avoids
the unnecessary steps for building an already shown menu.