[fenix] For https://github.com/mozilla-mobile/fenix/issues/4977: Migrate fennec shortcuts (https://github.com/mozilla-mobile/fenix/pull/7251)
* For https://github.com/mozilla-mobile/fenix/issues/4977 - Support opening Fennec pinned website shortcuts in Fenix

Fennec's pinned website shortcuts are set to open the BrowserApp activity.
So we need a new activity alias to actually catch such Intents. Otherwise they
would open "org.mozilla.firefox/.App" without any way to inform that this is
the result of the user clicking on a pinned shortcut.
For actually checking if the newly received Intent is of a Fennec pinned
shortcut we introduce a new FennecBookmarkShortcutsIntentProcessor which will
prepare the Intent to open the shortcut's URL in a new tab.

* For https://github.com/mozilla-mobile/fenix/issues/4977 - Don't keep IntentReceiverActivity on the back stack

For successive Fennec pinned shortcuts to create a new IntentReceiverActivity
and be processed as normal we need to not keep this as our task root.

* For https://github.com/mozilla-mobile/fenix/issues/4977 - Test the FennecBookmarkShortcutsIntentProcessor