[components] Fix for #6155 - Back button from migration complete notification routes back to migration screen
Tapping the migration complete notification would cause activities stack to be
polluted with the migration progress activity which means that popping an
activity from the onboarding screen would lead back to the migration progress
activity, rather than finishing the app instance.

Placing the FLAG_ACTIVITY_CLEAR_TOP flag to the migration complete
notification's intent will ensure us that, once in the onboarding screen, we
can never go back to the migration one, by clearing everything that is at the
top.