[components] For https://github.com/mozilla-mobile/android-components/issues/6978 - Catch reflection exceptions in BrowserGestureDetector
Although the fields BrowserGestureDetector want to modify through reflection
should be a stable part of the Android framework it was seen that on some
devices they do not exist.
Nothing more we can do in these cases than just prevent the reflection
exceptions bubbling up and potentially crash client apps.