[fenix] For https://github.com/mozilla-mobile/fenix/issues/11664 — Fixup MissingResourceExceptions being thrown in exotic Locales (https://github.com/mozilla-mobile/fenix/pull/13124)
Our kotlin code is not catching the `MissingResourceException` in the `LeanplumMetricsService` which results in the app crashing when the locale isn't known by the device.

Catches the exception, and falls back to the ISO 639 language code. This isn't a great solution, because ISO 639 isn't especially stable.

In practice however this is almost certainly never going to be a problem because Leanplum isn't going to be supported in such exotic locales.

In this case, using the ISO 639 language code allows the error message to be more informative.