[components] Closes https://github.com/mozilla-mobile/android-components/issues/8046: Remove unneeded unload handlers to make bfcache work
Window unload handlers block the bfcache. The extension framework
already automatically disconnects ports on unload so they aren't
needed in the first place.

For the reader mode content script, we are now connecting and
disconnecting on pageshow/pagehide to support readerable checks
when navigating (back/forward) to cached pages. We don't need to
do that for reader pages, as we opted out of the cache there.

For FxA there's no difference in behaviour. I've also tested
navigating back and forth during the signup flow and it still
works, as expected.