[focus] Close https://github.com/mozilla-mobile/focus-android/issues/4602: Move Custom Tab session ID check to MainActivity
The check that we originally added would usually have been enough, but
we made a mistake in on caling super. The problem gets complicated
there, because we want to call super for the `AppCompatActivity` and not
the `MainActivity` which is where the crash happens when we use a
session ID in the MainActivity for observing session changes that are
happening on a non-existent Custom Tab.

The fix here is to move the check we had in the `CustomTabActivity` to
the `MainActivity`. This is by no means ideal in term of
separation-of-concerns, but for a patch fix, it allows us to gracefully
exit.