[fenix] For https://github.com/mozilla-mobile/fenix/issues/14980 - Effectively disable tabs tray STATE_HALF_EXPANDED (https://github.com/mozilla-mobile/fenix/pull/16052)
STATE_HALF_EXPANDED cannot be disabled while also keeping fitToContents = true
based on which the tabs tray layout is currently set.
To work around this we'll set a a minuscule height for the tab tray when in
this state and then immediately advance to STATE_HIDDEN so to make it
imperceptible to the users.
Since I couldn't write unit tests because of InflateExceptions in Robolectric
I've written UI tests to protect against regressions.