[components] For https://github.com/mozilla-mobile/android-components/issues/9351 - Get camera image even with "Don't keep activities"
The previous functionality was already based on saving the camera photo in a
specific file and then us reading it from there but what that file name was was
lost in the case of the activity being destroyed as being the case with "Don't
keep activities".
By making that file name top level it will now be available for the entire
duration of the app.