[fenix] Closes https://github.com/mozilla-mobile/fenix/issues/18150: select tabs button not displayed (https://github.com/mozilla-mobile/fenix/pull/18216)
For https://github.com/mozilla-mobile/fenix/issues/18150: Improve showing/hiding of "Select tabs" button in tabs tray

Payload argument is unavailable after first call of notifyItemChanged() as the
whole View need to be updated.
Inject a "isPrivate" lambda in the adapter so it will always know what tabs are
currently showing and so always ensure the expected behavior.