[fenix] For https://github.com/mozilla-mobile/fenix/issues/17805 - Fix adjustResize deprecation (https://github.com/mozilla-mobile/fenix/pull/18252)
* For https://github.com/mozilla-mobile/fenix/issues/17805 - Fix adjustResize deprecation

To handle the deprecation of `adjustResize` I've moved it from `styles.xml` and `AndroidManifest.xml` to `Activity.kt` as a fallback for devices with Android < 11. For Android 11 and up `setDecorFitsSystemWindows(false)` and `OnApplyWindowInsetsListener` will be used to handle app insets. Normal use activities should call `enableSystemInsetsHandling` in `onCreate` as to properly display system bars and for proper keyboard handling.