[components] Fix https://github.com/mozilla-mobile/android-components/issues/9614 - Use the new InputResultDetail
Both NestedGeckoView and NestedWebView will now return an InputResultDetail
wrapping many new details about how a touch event will be handled.

NestedGeckoView's InputResultDetail will be used to more accurately decide when
to animate the toolbar or start the pull to refresh feature.

NestedWebView's InputResultDetail will only have details about if it will
handle the touch or not. With all the other being unknown the dynamic toolbar
or pull to refresh features will not work.