[components] For https://github.com/mozilla-mobile/android-components/issues/8989 - Don't catch/rethrow prompt cast exceptions
This try-catch was added in https://github.com/mozilla-mobile/android-components/issues/6687 as an effort to better individualize prompt
cast exceptions in our crash reporting platforms to better asses their
frequency and hopefully get more data to debug and fix the underlying issue.

The original issue is resolved in the first commit of this patch so there is no
need to keep the try-catch.