[focus] Closes https://github.com/mozilla-mobile/focus-android/issues/5134: remove unnecessary coroutines proguard rules.
According to the coroutines docs, these rules are bundled with the
library so we no longer need to add them to our project. These rules
were added in 7a1a2ed67318c1b5b34f9bb36d2da8d93dc35fb0 and this commit
is effectively a revert of that commit.